requirements.txt
astroid==2.4.2
certifi==2020.4.5.1
chardet==3.0.4
ffmpeg-python==0.2.0
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psutil==5.7.0
pylint==2.5.3
PyQRCode==1.2.1
python-engineio==3.12.1
python-magic==0.4.18
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
sockets==1.0.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.9
websocket-client==0.57.0
wrapt==1.12.1
